553864 Fixed GameState and all Unit classes. There were ints where there should have been doubles.

+++ AttackUnit.java
@@ -1,6 +1,6 @@
 public class AttackUnit extends Unit {
 
-   private int moveSpeed;
+   private double moveSpeed;
 
    /** Default AttackUnit constructor. */
    public AttackUnit() {
@@ -9,9 +9,9 @@ public class AttackUnit extends Unit {
    }
 
    /** Creates an AttackUnit with the given fields. */
-   public AttackUnit(int id, int posX, int posY, int maxHealth, 
-    int attackStrength, int defenseStrength, double attackSpeed, 
-    int attackRange, int moveSpeed) {
+   public AttackUnit(int id, int posX, int posY, double maxHealth, 
+    double attackStrength, double defenseStrength, double attackSpeed, 
+    double attackRange, double moveSpeed) {
 
       super(id, posX, posY, Unit.UnitType.Attack, maxHealth, attackStrength,
        defenseStrength, attackSpeed, attackRange); 
@@ -22,7 +22,7 @@ public class AttackUnit extends Unit {
     * Gets this AttackUnit's move speed.
     * @return The AttackUnit's move speed.
     */
-   public int getMoveSpeed() {
+   public double getMoveSpeed() {
       return moveSpeed;
    }
